Extract metrics signals (break circular dependency) and use proper tasks decorator, logging and imports
