+++ udata/core/metrics/signals.py
@@ -1,0 +1,12 @@
+# -*- coding: utf-8 -*-
+from __future__ import unicode_literals
+
+from flask.signals import Namespace
+
+ns = Namespace()
+
+#: Sent when a metric need to be updated
+metric_need_update = ns.signal('metric:need-update')
+
+#: Sent when a metric has been updated
+metric_updated = ns.signal('metric:updated')
